Miscellaneous compilation and warning fixes.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@31778 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/toolbar/toolbar.cpp
@@ -628,7 +628,8 @@ void MyFrame::OnToolLeftClick(wxCommandEvent& event)
 void MyFrame::OnToolRightClick(wxCommandEvent& event)
 {
     m_textWindow->AppendText(
-            wxString::Format(_T("Tool %d right clicked.\n"), event.GetInt()));
+            wxString::Format(_T("Tool %d right clicked.\n"),
+                             (int) event.GetInt()));
 }
 
 void MyFrame::OnCombo(wxCommandEvent& event)
